Fixes for Skia iOS swarming bot
BUG=skia:5155

Review URL: https://codereview.chromium.org/1859603002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299676 0039d316-1c4b-4281-b951-d872f2087c98

+++ scripts/slave/recipe_modules/skia/ios_flavor.py
@@ -109,16 +109,17 @@ class iOSFlavorUtils(default_flavor.DefaultFlavorUtils):
 
   def cleanup_steps(self):
     """Run any device-specific cleanup steps."""
-    self._skia_api.run(
-        self._skia_api.m.step,
-        name='reboot',
-        cmd=[self.ios_bin.join('ios_restart')],
-        infra_step=True)
-    self._skia_api.run(
-        self._skia_api.m.step,
-        name='wait for reboot',
-        cmd=['sleep', '20'],
-        infra_step=True)
+    if self._skia_api.do_test_steps or self._skia_api.do_perf_steps:
+      self._skia_api.run(
+          self._skia_api.m.step,
+          name='reboot',
+          cmd=[self.ios_bin.join('ios_restart')],
+          infra_step=True)
+      self._skia_api.run(
+          self._skia_api.m.step,
+          name='wait for reboot',
+          cmd=['sleep', '20'],
+          infra_step=True)
 
   def read_file_on_device(self, path):
     """Read the given file."""
